Use FindTime object, rectify geometry types
New FindTime object persists loaded data.

Geometry:
KDE can is some cases have multiple polygons, so the type has to be
Multipolygon.
Now geometries are converted into the proper type before DB insert.

Also, the task function doesn't determine whether it is in the
background, so removed the mention of it from the name